// src/aws/types/mediatypes.ts
// charset=UTF-8 is deprecated but mobile applications sometimes needs it
export enum MediaType {
  APPLICATION_JSON = "application/json",
  APPLICATION_XML = "application/xml",
  APPLICATION_GEOJSON = "application/geo+json;charset=UTF-8",
  IMAGE_SVG = "image/svg+xml",
  IMAGE_JPEG = "image/jpeg",
  TEXT_PLAIN = "text/plain",
  TEXT_HTML = "text/html",
  TEXT_CSV = "text/csv",
  APPLICATION_JSON_UTF8 = "application/json;charset=UTF-8",
}